scripts/requirements.txt
pyyaml
requests

